fix: Fix regex and test
